Add Terminal interface for consistent progress output
Terminal is an interface that describes a consistent terminal
printing experience. There is a built in ThrottleTerminal
type that throttles writing of progress messages, and
Printf and Errorf functions that know how to advance past a
progress line.

The existing prints in the code were changed to use the
Terminal interface instead of raw fmt.Printf and
log.Fatalf.

Cleaned up platform-specific code a little.

Fixed some bugs in commit walking, although this is in code
that is going to be completely replaced soon. It only existed
and was debugged to better understand the process of fetching
Git status info.