New faster commit fetching, better load/save
Starting to implement incremental vcs operations that are still
fast. RunExternalIncremental and RunGitCommandIncremental pass
lines from stdout and stderr to user callbacks, and this still
runs as fast as before.

Majorly refactored commit fetching. Now it starts to respect
saved data, so that the repo can be kept up to date as part
of normal operation.

Load/save cleaned up a lot. Tried to remove as many of the
previous bug points as possible, e.g. as little duplication of
data, using methods on small structs to avoid chance of copy/paste
errors going unnoticed, etc.